Added support for the openSUSE platform
Includes tests in test kitchen. Note: the mysql cookbook didn't support opensuse either, so instead of having to open yet another PR to another cookbook, I'm just using the postgresql cookbook (which i've opened a PR for opensuse support).

(As a further improvement, it might make sense to use a more lightweight db that doesn't assume too much of the system it's running on -- perhaps an in-memory db -- so that this cookbook could just focus on installing and testing flyway itself, not the DB).